Bug fixed. Commit by Coco 2020-8-9 12:36PM

+++ src/main/java/com/wlhse/service/FilePropagationPlanService.java
@@ -21,7 +21,7 @@ public interface FilePropagationPlanService {
 
     R insertNewFilePropagationDetail(List<FilePropagationDetailDto> filePropagationDetailDto);
 
-    R deleteFilePropagationPlan(int id);
+    R deleteFilePropagationPlan(Long id);
 
     R deleteFilePropagationPlanDetail(int id);
